# Add optional matcher to IndexReader.LabelValues