--- a/pkg/storegateway/indexheader/lazy_binary_reader.go
+++ b/pkg/storegateway/indexheader/lazy_binary_reader.go
@@ -188,7 +188,7 @@ func (r *LazyBinaryReader) LookupSymbol(o uint32) (string, error) {
 }
 
 // LabelValues implements Reader.
-func (r *LazyBinaryReader) LabelValues(name string) ([]string, error) {
+func (r *LazyBinaryReader) LabelValues(name string, filter func(string) bool) ([]string, error) {
 	r.readerMx.RLock()
 	defer r.readerMx.RUnlock()
 
@@ -197,7 +197,7 @@ func (r *LazyBinaryReader) LabelValues(name string) ([]string, error) {
 	}
 
 	r.usedAt.Store(time.Now().UnixNano())
-	return r.reader.LabelValues(name)
+	return r.reader.LabelValues(name, filter)
 }
 
 // LabelNames implements Reader.