@pstibrany 2022-11-07
> Makes sense to me. I also agree with Peter's comments.

My comments are non-blocking nits, feel free to ignore them.